fixed user deletion bug

+++ src/user/user.controller.ts
@@ -125,14 +125,15 @@ export class UserController {
       select: ['id'],
     })
 
-    const pids = _join(_map(user.projects, el => `'${el.id}'`), ',')
-    const query1 = `ALTER table analytics DELETE WHERE pid IN (${pids})`
-    const query2 = `ALTER table customEV DELETE WHERE pid IN (${pids})`
-
     try {
-      await this.projectService.deleteMultiple(pids)
-      await clickhouse.query(query1).toPromise()
-      await clickhouse.query(query2).toPromise()
+      if (!_isEmpty(user.projects)) {
+        const pids = _join(_map(user.projects, el => `'${el.id}'`), ',')
+        const query1 = `ALTER table analytics DELETE WHERE pid IN (${pids})`
+        const query2 = `ALTER table customEV DELETE WHERE pid IN (${pids})`
+        await this.projectService.deleteMultiple(pids)
+        await clickhouse.query(query1).toPromise()
+        await clickhouse.query(query2).toPromise()
+      }
       await this.userService.delete(id)
       
       return 'Account has been deleted'
